Clean up / add character-frame ilvl display, and pref for same
This commit finishes the existing work to show ilevel on items in the
character frame, adding an appropriate option checkbox to the prefs.

Along the way it does a little tidying and clarifying of the "items"
versus "nditems" contents, and scoops out some duplicate logic into
separate, more generic functions.

Also it bumps the TOC to 90002 and makes the appropriate code changes to
the detection of Classic vs Retail.